Make context menus more context sensitive
Context menus are now more contextually sensitive. Depending on if the
node you've right clicked on is "internal", "default", or neither, you
will get different options.

This was done because I didn't want to bother implementing ways to "see
inside" of the default nodes. This functionality may be added later,
after all the internal systems are done.