Start on Node execution scheduling.
Node execution is now scheduled to happen, rather than being called
immediately. The thought process behind this is that it will prevent
double executions, and will help keep node systems synchronized.

While it definitely will stop the former, we have to wait and see on the
latter.

This system could be improved later, eventually. At least if the current
one isn't good enough.