Start work on Node saving/loading
This commit adds functionality for saving and loading custom nodes, and
eventually entire projects.

This system produces Node files with the .node file extension. This is
purely a binary file, so the file extension is just for show, and maybe
to prevent loading of random file types. Regardless, the node is
converted to JSON, and compressed using LZString. The compressed data is
then saved as a ".node" file.